requirements.txt
boto==2.38.0
peewee==2.6.3
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
